+++ opa-datafilter-mongo-spring-boot-starter/src/main/java/com/github/jferrater/opadatafiltermongospringbootstarter/repository/OpaDataFilterMongoRepository.java
@@ -1,5 +1,7 @@
 package com.github.jferrater.opadatafiltermongospringbootstarter.repository;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.repository.NoRepositoryBean;
@@ -18,4 +20,6 @@ public interface OpaDataFilterMongoRepository<T, ID> extends MongoRepository<T, 
     List<T> findAll();
 
     List<T> findAll(Sort sort);
+
+    Page<T> findAll(Pageable pageable);
 }