+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MongoDB: findAll(Sort sort) implementation of OpaDataFilterMongoRepository
 - Core: Pretty json print log of Partial Request
 - JPA: findAll(Pageable pageable) implementation of OpaDatafilterRepository
+- MongoDB: findAll(Pageable pageable) implementation of OpaDataFilterMongoRepository
 
 ## [0.4.4] - 2020-04-21
 ### Added